requirements.txt
azure-storage-blob
bokeh==2.4.3
hydralit
keyboard
pillow
scikit-learn
matplotlib
seaborn
bioinfokit
tk
ppscore
st-click-detector
streamlit==1.11.1
dtale[streamlit]==2.8.1
streamlit-option-menu
streamlit-extras
streamlit-quill
streamlit-pandas-profiling
python-dotenv
watchdog